feat(api): dockerize application and implement robust module filtering
- Add Dockerfiles and docker-compose orchestration
- Implement defensive link resolution to fix AttributeError
- Add case-insensitive subject filtering for modules
- Support optional class assignment via kode_kelas in file upload
- Add mongo-init.js for automated database migration

+++ models/timestamp.py
@@ -4,14 +4,14 @@ from beanie import before_event, Insert, Replace, Save, SaveChanges, Document
 from pydantic import Field
 
 
-def current_timestamp() -> float:
-    return datetime.now(timezone.utc).timestamp()
+def current_timestamp() -> int:
+    return int(datetime.now(timezone.utc).timestamp())
 
 
 class TimestampDocument(Document):
-    created_at: float = Field(default_factory=current_timestamp)
-    updated_at: float = Field(default_factory=current_timestamp)
-    deleted_at: Optional[float] = None
+    created_at: int = Field(default_factory=current_timestamp)
+    updated_at: int = Field(default_factory=current_timestamp)
+    deleted_at: Optional[int] = None
 
     @before_event(Insert)
     def sync_timestamps(self):
